Add i18n translations on the analytics page
