+++ logicle/locales/en/common.json
@@ -238,6 +238,10 @@
   "administrator-settings": "Administrator settings",
   "dashboard": "Dashboard",
   "users": "Users",
+  "conversations": "Conversations",
+  "messages": "Messages",
+  "daily-usage": "Daily Usage",
+  "most-active-workspaces": "Most active workspaces",
   "workspaces": "Workspaces",
   "assistants": "Assistants",
   "backends": "Backends",
